added sign in functionality, create class functionality on frontend and corrosponding changed on backend

+++ prodigy_backend/classes/models.py
@@ -8,7 +8,7 @@ class Class(models.Model):
     subject = models.ForeignKey(Subject, on_delete=models.PROTECT, null=True)
     teacher = models.ForeignKey(Profile, on_delete=models.CASCADE)
     desc = models.CharField(max_length=800)
-    cost = models.PositiveSmallIntegerField()
+    cost = models.DecimalField(max_digits=8, decimal_places=2)
 
     def __str__(self):
         return f"{self.name} by {self.teacher.user.username} for ${self.cost}"
@@ -24,12 +24,12 @@ class Class(models.Model):
 
 class Review(models.Model):
     text = models.CharField(max_length=1000)
-    rating = models.IntegerField(
-        validators=[
-            MaxValueValidator(5),
-            MinValueValidator(0)
-        ]
-    )
+    rating = models.DecimalField(max_digits=2, decimal_places=1,
+                                 validators=[
+                                     MaxValueValidator(5),
+                                     MinValueValidator(0)
+                                 ]
+                                 )
     class_for = models.ForeignKey(Class, on_delete=models.CASCADE)
     user_for = models.ForeignKey(
         Profile, on_delete=models.CASCADE, related_name='user_reviewed')
